Fix SPE-2297 - slowdown after overhang.
The bug lies in the fact that SmoothPathCache comapare path attributes
in `resolve_or_fit_split_with_seam`. For some paths it does the
following: If attributes of paths are the same it merges the paths.

Since the == operator of ExtrusionAttributes did not consider
overhang_attributes in the comparison, some paths were
merged even though they had different overhang_attributes.
This led to the wrong overhang attributes being applied to
some paths.

Now the == operator of ExtrusionAttributes takes overhang_attributes
into account and it fixes the issue.